atc: behavior: don't allow unique resource types
this seems to unnecessarily complicate the codebase, and i don't think
many people use it or even know it exists or what it's for.

this field makes it so that, with global resources enabled, resources
using the type will have their own unique history, rather than being
global.

this is really just to support the time resource, so that all jobs don't
run on the same interval. but the time resource is built in, not a
custom type, so let's just reserve this behavior for built-in resource
types only until we eventually remove it.

note: i did a code search to see if i could find anyone using this in
the wild, and there was only result, and it was for the cron resource,
which doesn't benefit from unique history anyway (unlike the time
resource).

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>