atc: behavior: check step figures out config/scope
previously the check step was completely reliant on the delegate and
metadata to figure out the resource config and scope to save to.

now the check step constructs its own config and scope at runtime based
on the evaluated config. this makes the check step more independent, and
makes it easier to have multiple check steps in one build plan;
previously it would only save to the scope associated with the db.Check,
but now that it's decoupled it should easier to run multiple in a
db.Build.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>