atc: behavior: push logic down to check step
determine cursor version in check step based on current scope, rather
than determining the scope and cursor version on the outside.

now that the scope isn't determined until the check step runs, this
means that we will no longer be able to de-dupe checks for the same
resource config/scope when queueing up checks.

as-is, this will be pretty wasteful: we'll may end up queueing a bunch
of redundant checks for the same config/scope, pretty much undoing the
benefits of global resources.

the next step is to mitigated this by having the 'check' step to skip
the actual checking if it notices that the config/scope in question has
already been checked within the interval. to do this, we'll need to pass
the interval to the check step. we can still use the interval on the
outside to avoid queueing for resources which have already recently been
checked recently.

with that mitigation, there will still be the overhead of queueing and
running the build itself, but at least the build will still do something
useful: the check step will update the config/scope on the resource or
resource type, which is still necessary to do even with global
resources.

the biggest advantage of this approach is being able to simplify the
architecture. the check step becomes the source of truth, and can be
injected into build plans with less effort, and be trusted to handle the
cursor version, updating the resource/resource type, and doing a no-op
if its scope has already been checked (once that's been implemented,
anyway).

note: now that the check factory (outside) doesn't create a scope, it
can't possibly create a db.Check, so that's all been commented out for
now. the next step here is to change it to create a db.Build.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>