atc: behavior: check factory creates builds
as-is, this will be pretty wasteful: it'll end up queueing a bunch of
redundant checks for resources with the same config/scope, and they'll
all run their checks, pretty much undoing the benefits of global
resources.

the next step is to mitigate this by having the 'check' step regulate
itself by skipping the check if its interval has not elapsed, while
locking to prevent concurrent checks of the same scope.

with that mitigation, there will still be the overhead of queueing and
builds for all resources for each scope to check, but at least the build
will do something useful: its check step will update the config/scope on
the resource or resource type.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>