atc: behavior: grace period for resource config gc
previously, when the GC ran in the period of time between a config being
constructed and being assigned to its resource (which includes running
the check), the config would be removed, causing mayhem in the check
step.

now, resource configs will be marked as 'referenced' by the check step,
and they will be removed when they have not been referenced for a while
(and are not referenced by a resource, resource type, or resource cache
by that point).

"a while" is hardcoded to the overall check timeout + spare change (5
minutes). it honestly doesn't matter too much because they're literally
just rows in the database at that point (caches go away before configs),
so hours is better than seconds here; don't want it to happen during a
slow check, because it won't be assigned to the resource by then.

one nifty thing: this effectively removes the need to use
resource_config_check_sessions to keep resource configs alive, so this
simplifies the GC query. i haven't benchmarked or EXPLAIN ANAYLZEd it
though.

...and another thing: it may be tempting to simplify the resource config
query completely down to just use the 'last referenced' time, because
their periodic checking effectively heartbeats it, but simplifying it
that far is a little risky. it's only bumped when a check runs, so if
the check_every is longer than the grace period (maybe they're using
webhooks), the config might get GC'd. i suppose you could just bump this
default value even higher then.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>