atc: behavior: don't assume builds have jobs
previously, many areas of the codebase - primary logging and tracing -
assumed that builds had a job name.

this commit centralizes the construction of lager.Data and tracing.Attrs
for builds onto the db.Build type itself.

it also modifies the step tracing flow to go through the step delegate.
this is ultimately so that we can call the new .TracingAttrs method on
the build, but also makes it a bit easier to test.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>