atc: behavior: auth doesn't assume build has job
this used to call .Job with an empty job name, which was kind of weird.
instead, let's just say any pipeline build that has no job is treated
as 'private'.

this is consistent with the previous behavior of who could see resource
checking errors (only those authorized with the team).

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>